Create shared metadata section component
